+++ examples/data-sources/iosxe_interface_vrrp_v2/data-source.tf
@@ -1,0 +1,5 @@
+data "iosxe_interface_vrrp_v2" "example" {
+  type     = "Vlan"
+  name     = "123"
+  group_id = 1
+}